requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.9.0
scrapy==2.11.0
pymongo==4.6.0
elasticsearch==8.11.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml>=4.9.0
python-dotenv==1.0.0
eventlet==0.33.3
